fix otp code

+++ volumes/app/common/otp_manager.py
@@ -81,9 +81,12 @@ class OTPManager:
     def verify_otpcode(phone_number, code):
         try:
             verification_code = OTPCode.objects.get(phone_number=phone_number)
-            if int(code) == int(verification_code.code):
-                verification_code.delete()
-                return True
+            if code and code.isdigit():
+                if int(code) == int(verification_code.code):
+                    verification_code.delete()
+                    return True
+                else:
+                    return False
             else:
                 return False
         except ObjectDoesNotExist:
